attackRatio 적용, hp와 atk를 int에서 float로 바꾼다.

+++ VampireSurvivors/app/src/main/java/kr/ac/tukorea/www/smgp2018184001/vampiresurvivors/framework/interfaces/IAttackable.java
@@ -1,6 +1,6 @@
 package kr.ac.tukorea.www.smgp2018184001.vampiresurvivors.framework.interfaces;
 
 public interface IAttackable {
-    public int getAtk();
+    public float getAtk();
     public boolean isAttacking();
 }
