+++ VampireSurvivors/app/src/main/java/kr/ac/tukorea/www/smgp2018184001/vampiresurvivors/game/weapon/Lightning.java
@@ -42,7 +42,7 @@ public class Lightning extends Object implements IAttackable {
     }
 
     @Override
-    public int getAtk() {
+    public float getAtk() {
         return 0;
     }
